requirements.txt
gym==0.10.11
imageio==2.4.0
IPython
matplotlib
PILLOW
pyglet==1.3.2
pyvirtualdisplay
tensorflow
tf-agents-nightly
xvfbwrapper
